test DuolionMetric only with big graph
